Style updates to scram sasl support
